requirements.txt
docx==0.2.4
docx2pdf==0.1.8
httpx==0.28.1
mammoth==1.10.0
markdownify==1.2.0
markitdown==0.1.2
MarkupSafe==3.0.2
openai==1.99.9
openai-whisper==20250625
opencv-python==4.12.0.88
python-docx==1.2.0